refactor(engine-core): split contracts/analyze/cli modules
