+++ .pr-body-rust-engine-core-modularization-v1.md
@@ -1,0 +1,26 @@
+## Summary
+- Refactor `crates/engine-core` into clear modules while preserving behavior and API shape:
+  - `contract.rs`: all request/response/IR/diagnostic contract types
+  - `analyze.rs`: core `analyze` flow
+  - `normalize.rs`: internal normalization helper(s)
+  - `cli.rs`: binary command execution flow
+  - `error.rs`: CLI error-to-exit mapping
+- Keep `lib.rs` and `main.rs` as crate-local glue/re-export boundaries only.
+- No functional changes to analyze output, CLI usage string, or exit code behavior.
+
+## TDD
+- Kept existing tests intact; no test logic changes.
+- Verified existing contract round-trip and analyze mapping tests continue to pass unchanged.
+
+## Validation
+- Full gate run passed in CI order:
+  1. `pnpm run lint`
+  2. `pnpm run format:check`
+  3. `pnpm run build`
+  4. `pnpm run test`
+  5. `cargo fmt --all --check`
+  6. `cargo clippy --workspace --all-targets -- -D warnings`
+  7. `cargo test --workspace --all-targets`
+
+## Notes
+- `guard:runtime` script is not present on current `main`; validation follows current `.github/workflows/ci.yml` order exactly.
