+++ crates/engine-core/src/main.rs
@@ -1,28 +1,13 @@
-use std::io::{self, Read};
-
-use engine_core::{analyze, AnalyzeRequest};
+mod cli;
+mod error;
 
 fn main() {
-    let mut args = std::env::args();
-    let _bin = args.next();
-
-    match args.next().as_deref() {
-        Some("analyze") => {
-            let mut input = String::new();
-            io::stdin()
-                .read_to_string(&mut input)
-                .expect("failed to read stdin");
-
-            let request: AnalyzeRequest =
-                serde_json::from_str(&input).expect("failed to parse AnalyzeRequest JSON");
-            let response = analyze(request);
-            let output =
-                serde_json::to_string_pretty(&response).expect("failed to encode response");
-            println!("{output}");
+    if let Err(error) = cli::run() {
+        match error {
+            error::CliError::Usage => {
+                eprintln!("usage: engine-core analyze");
+            }
         }
-        _ => {
-            eprintln!("usage: engine-core analyze");
-            std::process::exit(2);
-        }
+        std::process::exit(error.exit_code());
     }
 }
